refactor: reorganize package structure for Swing components and update imports
